+++ flutter_secure_storage/CHANGELOG.md
@@ -1,3 +1,9 @@
+## Fork
+
+* Enabled StrongBox by default, use fallback if it's not available.
+* [Android] Allow to force StrongBox with a flag (onlyAllowStrongBox)
+* [Android] Method to check if an Android device supports Strongbox
+
 ## 10.0.0
 This major release brings significant security improvements, platform updates, and modernization across all supported platforms.